Move StarknetFeePayment and StarknetPriceUnit to better locations
