+++ Sources/Starknet/Data/PriceUnit.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+public enum StarknetPriceUnit: String, Codable {
+    case wei = "WEI"
+    case fri = "FRI"
+}
