+++ Sources/Starknet/Data/Transaction/Data/FeePayment.swift
@@ -1,0 +1,11 @@
+import Foundation
+
+public struct StarknetFeePayment: Decodable, Equatable {
+    public let amount: Felt
+    public let feeUnit: StarknetPriceUnit
+
+    enum CodingKeys: String, CodingKey {
+        case amount
+        case feeUnit = "unit"
+    }
+}
